CloudFunctions/fetch-latest-emails-pubsub/requirements.txt
functions-framework==3.*
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-firestore
google-cloud-aiplatform
google-cloud-pubsub==2.21.5